+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 ## [v1.4.0 - 2023-02-21]
 ### New
 - add maillib module
-- add ldap module
+- add ldaplib module
 - add ExecuteOsCommand to common
 - move tests to there modules
 - add workflows for sub directories